+++ src/Boot/src/BootloadManager/BootloadManager.php
@@ -4,45 +4,33 @@ declare(strict_types=1);
 
 namespace Spiral\Boot\BootloadManager;
 
-use Closure;
-use Spiral\Boot\Bootloader\BootloaderInterface;
-use Spiral\Boot\BootloadManagerInterface;
-use Spiral\Core\Container;
 use Spiral\Core\InvokerInterface;
 use Spiral\Core\ResolverInterface;
 use Spiral\Core\ScopeInterface;
 
 /**
- * Provides ability to bootload ServiceProviders.
+ * @deprecated since v3.4. Use the {@see StrategyBasedBootloadManager} instead.
  */
-final class BootloadManager implements BootloadManagerInterface, Container\SingletonInterface
+final class BootloadManager extends AbstractBootloadManager
 {
+    private InvokerStrategyInterface $invokerStrategy;
+
     public function __construct(
-        /* @internal */
-        private readonly ScopeInterface $scope,
+        ScopeInterface $scope,
         private readonly InvokerInterface $invoker,
         private readonly ResolverInterface $resolver,
-        private readonly Initializer $initializer
+        InitializerInterface $initializer,
+        ?InvokerStrategyInterface $invokerStrategy = null
     ) {
-    }
+        parent::__construct($scope, $initializer);
 
-    public function getClasses(): array
-    {
-        return $this->initializer->getRegistry()->getClasses();
-    }
-
-    public function bootload(array $classes, array $bootingCallbacks = [], array $bootedCallbacks = []): void
-    {
-        $this->scope->runScope(
-            [self::class => $this],
-            function () use ($classes, $bootingCallbacks, $bootedCallbacks): void {
-                $this->boot($classes, $bootingCallbacks, $bootedCallbacks);
-            }
-        );
+        $this->invokerStrategy = $invokerStrategy ?? new DefaultInvokerStrategy(...$this->resolver->resolveArguments(
+            (new \ReflectionClass(DefaultInvokerStrategy::class))->getConstructor()
+        ));
     }
 
     /**
-     * Bootloader all given classes.
+     * Bootload all given bootloaders.
      *
      * @param array<class-string>|array<class-string, array<string,mixed>> $classes
      *
@@ -50,45 +38,6 @@ final class BootloadManager implements BootloadManagerInterface, Container\Singl
      */
     protected function boot(array $classes, array $bootingCallbacks, array $bootedCallbacks): void
     {
-        $bootloaders = \iterator_to_array($this->initializer->init($classes));
-
-        foreach ($bootloaders as $data) {
-            $this->invokeBootloader($data['bootloader'], Methods::INIT, $data['options']);
-        }
-
-        $this->fireCallbacks($bootingCallbacks);
-
-        foreach ($bootloaders as $data) {
-            $this->invokeBootloader($data['bootloader'], Methods::BOOT, $data['options']);
-        }
-
-        $this->fireCallbacks($bootedCallbacks);
-    }
-
-    private function invokeBootloader(BootloaderInterface $bootloader, Methods $method, array $options): void
-    {
-        $refl = new \ReflectionClass($bootloader);
-        if (!$refl->hasMethod($method->value)) {
-            return;
-        }
-
-        $method = $refl->getMethod($method->value);
-
-        $args = $this->resolver->resolveArguments($method);
-        if (!isset($args['boot'])) {
-            $args['boot'] = $options;
-        }
-
-        $method->invokeArgs($bootloader, \array_values($args));
-    }
-
-    /**
-     * @param array<Closure> $callbacks
-     */
-    private function fireCallbacks(array $callbacks): void
-    {
-        foreach ($callbacks as $callback) {
-            $this->invoker->invoke($callback);
-        }
+        $this->invokerStrategy->invokeBootloaders($classes, $bootingCallbacks, $bootedCallbacks);
     }
 }